requirements.txt
# Databricks Apps Requirements

# Core
streamlit>=1.28.0
pandas>=2.0.0

# Databricks SDK
databricks-sdk>=0.23.0
databricks-sql-connector>=3.0.0

# Visualization
plotly>=5.17.0
kaleido>=0.2.1

# Report Generation
reportlab>=4.0.0
jinja2>=3.1.0

# AWS Bedrock (optional)
boto3>=1.34.0
